+++ src/styled.js
@@ -5,8 +5,9 @@ import flatten from './flatten';
 import { ThemeProvider, withTheme } from './theme';
 
 const styled = (Comp, { name, multi, ...opts } = {}) => style => {
-  const Styled = React.forwardRef(({ children, ...props }, ref) => {
-    const { attrs, comp, child } = opts;
+  const Styled = React.forwardRef(({ childRef, ...props }, ref) => {
+    const { attrs, comp, child, childProps } = opts;
+    const { children } = props;
 
     const styleProps = multi ? style : { style };
     const styleKeys = Object.keys(styleProps);
@@ -24,15 +25,27 @@ const styled = (Comp, { name, multi, ...opts } = {}) => style => {
       ]);
     }
 
+    const parentProps = {
+      ...(attrs && attrs(props)),
+      ...props,
+      ...styles
+    };
+
     return React.createElement(
       comp || Comp,
-      {
-        ref,
-        ...(attrs && attrs(props)),
-        ...props,
-        ...styles
-      },
-      child ? React.createElement(child, {}, children) : children
+      { ref, ...parentProps },
+      child
+        ? React.createElement(
+            child,
+            {
+              ref: childRef,
+              ...(typeof childProps === 'function'
+                ? childProps(parentProps)
+                : childProps)
+            },
+            children
+          )
+        : children
     );
   });
 
@@ -52,7 +65,10 @@ const styled = (Comp, { name, multi, ...opts } = {}) => style => {
   };
 
   Styled.withComponent = comp => styled(Styled, { comp })(style);
-  Styled.withChild = child => styled(Styled, { name, child })();
+
+  Styled.withChild = (child, childProps) =>
+    styled(Styled, { child, childProps })();
+
   Styled.withTheme = () => withTheme(Styled);
 
   return Styled;